Update and rename README.md to index.html
# brick-breaker
Brick Breaker is a classic arcade-style game where the player controls a paddle to bounce a ball and break all the bricks on the screen. The main goal is to clear every brick without letting the ball fall below the paddle. Players must use skill, timing, and quick reflexes to succeed.